Fixed issue where client generating level would add n-1 planets to the scene

+++ include/DbgImgui.hpp
@@ -159,7 +159,7 @@ public:
 		style.WindowBorderSize = 1;
 		style.ChildBorderSize = 1;
 		style.PopupBorderSize = 1;
-		style.FrameBorderSize = is3D;
+		style.FrameBorderSize = (float)is3D;
 
 		style.WindowRounding = 3;
 		style.ChildRounding = 3;
